Move AppDialogButton component to elements dir
